+++ packages/integration-sdk/CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to
 - Added `getEntity({ _type, _key })` function
 - Added optional `sourceEntityType` property on `RelationshipMapping`
 
+### Changed
+
+- Deprecated `createIntegrationRelationship`. Developers should use the exported `createDirectRelationship` or `createMappedRelationship` functions.
+
 ## 2.5.0 - 2020-07-28
 
 ### Added